Add custom exception handling for note not found (404)

+++ src/main/java/com/sarth/notesapp/service/NoteService.java
@@ -1,5 +1,6 @@
 package com.sarth.notesapp.service;
 
+import com.sarth.notesapp.exception.NoteNotFoundException;
 import com.sarth.notesapp.model.Note;
 import com.sarth.notesapp.repository.NoteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,18 +24,21 @@ public class NoteService {
     }
 
     public Note getNoteById(Long id){
-        return noteRepository.findById(id).orElse(null);
+        return noteRepository.findById(id).orElseThrow(() -> new NoteNotFoundException("Note with id " + id + " not found"));
     }
 
     public void deleteNote(Long id){
+
+        if(!noteRepository.existsById(id)){
+            throw new NoteNotFoundException("Note with id " + id + " not found");
+        }
         noteRepository.deleteById(id);
     }
 
     public Note updateNote(Long id, Note updatedNote){
-        Note existingNote = noteRepository.findById(id).orElse(null);
-        if(existingNote == null){
-            return null;
-        }
+        Note existingNote = noteRepository.findById(id).orElseThrow(()
+                -> new NoteNotFoundException("Note with id " + id + " not found"));
+
         existingNote.setTitle(updatedNote.getTitle());
         existingNote.setContent(updatedNote.getContent());
         return noteRepository.save(existingNote);
